Change checks for className being something like "sfrac nowrap" to use "classList.contains('sfrac') because that's all that matters.
Making that change means that the CheckXXX functions that do that need to be passed elements, not text nodes, so a test is added for that (probably speeds things up a little).

Found some cases where wikipedia is using an image for inline stuff and that is via 'style' and 'link' elements. However, there is some span that can be used. Need to skip over the style/link to get to it.